template: Move card logic to a mixin
We are going to have multiple card variants, so we can't overload the
Card component with v-if/v-else conditionals. Next commits will create
other card components using the same mixin.

https://phabricator.endlessm.com/T31724
