+++ template-ui/src/components/Card.vue
@@ -21,57 +21,14 @@
 </template>
 
 <script>
-import { getThumbnail, goToContent } from 'kolibri-api';
-import { mapState, mapGetters } from 'vuex';
+import { mapState } from 'vuex';
+import cardMixin from '@/components/mixins/cardMixin';
 
 export default {
   props: ['node'],
-  data() {
-    return {
-      thumbnail: null,
-      isHovered: false,
-    };
-  },
+  mixins: [cardMixin],
   computed: {
     ...mapState(['isHighQualityMedia']),
-    ...mapGetters(['getCardLabel', 'getAsset']),
-    title() {
-      if (this.node.kind === 'topic') {
-        return this.getCardLabel(this.node);
-      }
-      return this.node.title;
-    },
-    isLeaf() {
-      return this.node.kind !== 'topic';
-    },
-  },
-  methods: {
-    goToContent,
-    getTopicUrl(node) {
-      return `/${node.id}`;
-    },
-    async getThumbnail() {
-      if (!this.node.thumbnail && process.env.VUE_APP_USE_MOCK_DATA === 'true') {
-        this.thumbnail = this.getAsset('defaultThumbnail');
-        return;
-      }
-      if (this.node.thumbnail) {
-        this.thumbnail = this.node.thumbnail;
-        return;
-      }
-      const thumbnail = await getThumbnail(this.node);
-      if (thumbnail) {
-        this.thumbnail = thumbnail;
-      } else {
-        this.thumbnail = this.getAsset('defaultThumbnail');
-      }
-    },
-    handleHover(hovered) {
-      this.isHovered = hovered;
-    },
-  },
-  created() {
-    this.getThumbnail();
   },
 };
 </script>
